backend/requirements-dev.txt
# Development Dependencies
# Include all production requirements
-r requirements.txt

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.24.1

# Development Tools
black==24.4.2
isort==5.13.2
flake8==6.0.0
mypy==1.5.1

# Documentation
mkdocs==1.5.2
mkdocs-material==9.1.21

# Debugging
pdbpp==0.10.3
ipython==8.14.0

# Environment Management
python-dotenv==1.0.1

# Code Quality
pre-commit==3.3.3
